Fixed tests, renamed filter file & removed class member name prefix constant
