+++ constants.py
@@ -10,8 +10,6 @@ SAMPLE_CODE_FOLDER_NAME = "examples"
 
 TRACE_DATA_FILE_ENDING = ".pytype"
 
-CLASS_MEMBER_NAME_PREFIX = "class."
-
 class TraceData:
     FILENAME = "Filename"
     CLASS = "Class"
